Updated to .NET Framework 4.5 and added ability for the installer to download the framework as part of the install. Changed to using ValueInjecter to do preset cloning rather than the fragile per-property method.

+++ VidCoderWorker/HandBrakeEncoder.cs
@@ -59,7 +59,7 @@ namespace VidCoderWorker
 					long affinityMask = 0x0;
 					for (int i = 0; i < coresToUse; i++)
 					{
-						affinityMask |= (1 << i);
+						affinityMask |= (uint)(1 << i);
 					}
 
 					process.ProcessorAffinity = (IntPtr) affinityMask;
